requirements.txt
django-5.2.8 == pip install django
django_environ-0.12.0 == pip install django-environ
pymysql-1.1.2 == pip install pymysql
pip install --upgrade pip setuptools wheel
Pillow-12.0.0 == pip install Pillow
fastapi-0.123.0 == pip install fastapi
python_jose-3.5.0 == pip install python-jose